requirements.txt
Django==3.1
djangorestframework==3.11.1
mysqlclient==1.4.6
mysql-connector==2.2.9